+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [1.4.0] - 2020
+## [1.4.0] - 2020-04-20
 
 ### Changes
 * add page_limit parameter to ReportBro constructor to define custom page limit